Refactor storage interfaces
