+++ core/blockchain/impl/storage_util.hpp
@@ -9,7 +9,7 @@
 #include "common/buffer.hpp"
 #include "primitives/block_header.hpp"
 #include "primitives/block_id.hpp"
-#include "storage/face/persistent_map.hpp"
+#include "storage/face/batchable.hpp"
 
 /**
  * Auxiliary functions to simplify usage of persistant map based storage
@@ -63,7 +63,7 @@ namespace kagome::blockchain {
    * @return storage error if any
    */
   outcome::result<void> putWithPrefix(
-      storage::face::PersistentMap<common::Buffer, common::Buffer> &map,
+      storage::face::Batchable<common::Buffer, common::Buffer> &map,
       prefix::Prefix prefix,
       primitives::BlockNumber num,
       common::Hash256 block_hash,
@@ -77,7 +77,7 @@ namespace kagome::blockchain {
    * @return encoded entry or error
    */
   outcome::result<common::Buffer> getWithPrefix(
-      const storage::face::PersistentMap<common::Buffer, common::Buffer> &map,
+      const storage::face::Batchable<common::Buffer, common::Buffer> &map,
       prefix::Prefix prefix,
       const primitives::BlockId &block_id);
