+++ core/storage/face/generic_storage.hpp
@@ -1,0 +1,9 @@
+/**
+ * Copyright Soramitsu Co., Ltd. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#ifndef KAGOME_GENERIC_STORAGE_HPP
+#define KAGOME_GENERIC_STORAGE_HPP
+
+#endif //KAGOME_GENERIC_STORAGE_HPP
